+++ projects/tests/dbs/sql-stuff/sub-query/sub-query.spec.ts
@@ -488,12 +488,10 @@ describe('sql-relations', () => {
           "orders": [
             Order {
               "amount": 10,
-              "customer": undefined,
               "id": 1,
             },
             Order {
               "amount": 15,
-              "customer": undefined,
               "id": 2,
             },
           ],
@@ -503,17 +501,14 @@ describe('sql-relations', () => {
           "orders": [
             Order {
               "amount": 40,
-              "customer": undefined,
               "id": 3,
             },
             Order {
               "amount": 5,
-              "customer": undefined,
               "id": 4,
             },
             Order {
               "amount": 7,
-              "customer": undefined,
               "id": 5,
             },
           ],
@@ -523,12 +518,10 @@ describe('sql-relations', () => {
           "orders": [
             Order {
               "amount": 90,
-              "customer": undefined,
               "id": 6,
             },
             Order {
               "amount": 3,
-              "customer": undefined,
               "id": 7,
             },
           ],