+++ projects/tests/tests/relations/test-backend-instance-method.spec.ts
@@ -58,8 +58,6 @@ describe('relations to one behavior', () => {
     Task.defaultCategoryId = 20
     expect(await serverRemult.repo(Task).findFirst({})).toMatchInlineSnapshot(`
       Task {
-        "cat2": null,
-        "category": undefined,
         "categoryId": 10,
         "id": 1,
         "title": "task 1",
@@ -68,8 +66,6 @@ describe('relations to one behavior', () => {
     let task = await r(Task).findFirst({})
     expect(task).toMatchInlineSnapshot(`
       Task {
-        "cat2": null,
-        "category": undefined,
         "id": 1,
         "title": "task 1",
       }
@@ -77,8 +73,6 @@ describe('relations to one behavior', () => {
     doOnServer = async (task) => {
       expect(task).toMatchInlineSnapshot(`
         Task {
-          "cat2": null,
-          "category": undefined,
           "categoryId": 10,
           "id": 1,
           "title": "task 12",
@@ -87,8 +81,6 @@ describe('relations to one behavior', () => {
       await task.save()
       expect(task).toMatchInlineSnapshot(`
         Task {
-          "cat2": null,
-          "category": undefined,
           "categoryId": 10,
           "id": 1,
           "title": "task 12",
@@ -98,8 +90,6 @@ describe('relations to one behavior', () => {
     await task!.something()
     expect(await serverRemult.repo(Task).findFirst({})).toMatchInlineSnapshot(`
       Task {
-        "cat2": null,
-        "category": undefined,
         "categoryId": 10,
         "id": 1,
         "title": "task 12",
@@ -107,8 +97,6 @@ describe('relations to one behavior', () => {
     `)
     expect(task).toMatchInlineSnapshot(`
       Task {
-        "cat2": null,
-        "category": undefined,
         "id": 1,
         "title": "task 12",
       }
@@ -116,8 +104,6 @@ describe('relations to one behavior', () => {
     doOnServer = async (task) => {
       expect(task).toMatchInlineSnapshot(`
         Task {
-          "cat2": null,
-          "category": undefined,
           "categoryId": 10,
           "id": 1,
           "title": "task 122",
@@ -126,8 +112,6 @@ describe('relations to one behavior', () => {
       await task.save()
       expect(task).toMatchInlineSnapshot(`
         Task {
-          "cat2": null,
-          "category": undefined,
           "categoryId": 10,
           "id": 1,
           "title": "task 122",
@@ -137,8 +121,6 @@ describe('relations to one behavior', () => {
     await task!.something()
     expect(await serverRemult.repo(Task).findFirst({})).toMatchInlineSnapshot(`
       Task {
-        "cat2": null,
-        "category": undefined,
         "categoryId": 10,
         "id": 1,
         "title": "task 122",
@@ -146,8 +128,6 @@ describe('relations to one behavior', () => {
     `)
     expect(task).toMatchInlineSnapshot(`
       Task {
-        "cat2": null,
-        "category": undefined,
         "id": 1,
         "title": "task 122",
       }